fix: resolve paths
