--- src/robots.config.ts
@@ -1,3 +1,0 @@
-import { rules } from '#robots-rules'
-
-export default rules
